Fix bug in build.py deps detection
This fixes an insufficient test for whether Aspen is already installed.
The text 'found' shows up in the traceback when
pkg_resources.get_distribution fails in aspen/__init__.py, and that's
the way in which build.py's `import aspen` fails once the Response
import in __init__.py is removed.

The fix here is to delegate to pip to check whether this or that library
is already installed, rather than trying to duplicate that functionality
ourselves.