Upgrade 'repr' to a 'backslashreplace' override
I wanted to get rid of `utils`, since that is a junkdrawer, a failure to
organize properly. The last remaining item was the 'repr' codec error
handling strategy, which we use a little bit in Aspen, but even moreso
in Pando, as part of error messaging for malformed configuration values.
Since Python 3.5 adds support for what we need, it seemed best to
reintroduce a `backcompat` module. And since Python 3.5 *extends* the
functionality of the old 'backslashreplace' strategy, it seemed safe to
*replace* that strategy rather than registering a separate one. Any code
outside of Aspen that assumes the old behavior will not be disappointed,
because the old behavior is a limited case of the new behavior. If any
code outside of Aspen *also* upgrades to the new behavior, we *could*
end up with an extra call or two (possibly worse?). Therefore, we bail
in that case: we only upgrade if no-one else has.